+++ command/trust/trust.go
@@ -7,6 +7,7 @@ import (
 	"io"
 	"io/ioutil"
 	"os"
+	"path/filepath"
 	"runtime"
 
 	"github.com/docker/docker/api/types"
@@ -33,7 +34,7 @@ const (
 // NewCommand returns `trust` to retrieve the certificates from the nitro proxy and install on the
 // host machine. The CA is used to sign certificates for websites and adding the certificate
 // to the system allows TLS connections to be considered valid and trusted from the container.
-func NewCommand(docker client.CommonAPIClient, output terminal.Outputer) *cobra.Command {
+func NewCommand(home string, docker client.CommonAPIClient, output terminal.Outputer) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:     "trust",
 		Short:   "Trust certificates for environment",
@@ -127,6 +128,18 @@ func NewCommand(docker client.CommonAPIClient, output terminal.Outputer) *cobra.
 			}
 
 			output.Done()
+
+			// copy the certificate into the nitro dir
+			cert, err := os.Create(filepath.Join(home, ".nitro", "nitro.crt"))
+			if err != nil {
+				return err
+			}
+			defer cert.Close()
+
+			// copy the contents
+			if _, err := io.Copy(cert, temp); err != nil {
+				return err
+			}
 
 			output.Info("Installing certificate (you might be prompted for your password)")